+++ internal/slices/minmax.go
@@ -1,7 +1,7 @@
 package slices
 
-func Min(slice []int) int {
-	var min int = slice[0]
+func Min(slice []float64) float64 {
+	var min float64 = slice[0]
 
 	for _, item := range slice[1:] {
 		if item < min {
@@ -12,13 +12,13 @@ func Min(slice []int) int {
 	return min
 }
 
-func Max(slice []int) int {
-	var max int = slice[0]
+func Max(slice []float64) float64 {
+	var max float64 = slice[0]
 
 	for _, item := range slice[1:] {
 		if item > max {
 			max = item
-		}	
+		}
 	}
 
 	return max
